Split nfapi_vnf_lib into common, LTE and NR static libraries
