+++ nfapi/oai_integration/CMakeLists.txt
@@ -33,7 +33,7 @@ target_include_directories(nfapi_user_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 add_library(nfapi_lte_user_lib
         vnf/nfapi_lte_vnf.c)
 target_link_libraries(nfapi_lte_user_lib PRIVATE asn1_lte_rrc_hdrs UTIL)
-target_link_libraries(nfapi_lte_user_lib PRIVATE nfapi_user_lib)
+target_link_libraries(nfapi_lte_user_lib PUBLIC nfapi_user_lib nfapi_lte_vnf_lib)
 
 # NR VNF OAI integration
 add_library(nfapi_nr_user_lib
@@ -47,7 +47,7 @@ else()
     target_compile_definitions(nfapi_nr_user_lib PRIVATE ENABLE_SOCKET)
 endif()
 target_link_libraries(nfapi_nr_user_lib PRIVATE asn1_nr_rrc_hdrs UTIL nr_fapi_p7)
-target_link_libraries(nfapi_nr_user_lib PRIVATE nfapi_user_lib)
+target_link_libraries(nfapi_nr_user_lib PUBLIC nfapi_user_lib nfapi_nr_vnf_lib)
 if(OAI_WLS)
     target_link_libraries(nfapi_nr_user_lib PRIVATE wls_integration_lib)
 endif()
