+++ nfapi/oai_integration/socket/CMakeLists.txt
@@ -17,4 +17,4 @@ target_link_libraries(nfapi_socket_lib PRIVATE nfapi_nr_lib nfapi_nr_user_lib)
 target_link_libraries(nfapi_socket_lib PRIVATE log_headers)
 target_link_libraries(nfapi_socket_lib PRIVATE asn1_lte_rrc_hdrs asn1_nr_rrc_hdrs)
 target_link_libraries(nfapi_socket_lib PRIVATE nr_fapi_p5 nr_fapi_p7)
-target_link_libraries(nfapi_socket_lib PRIVATE nfapi_vnf_lib nfapi_pnf_lib)
+target_link_libraries(nfapi_socket_lib PRIVATE nfapi_nr_vnf_lib nfapi_pnf_lib)
