+++ nfapi/oai_integration/wls_integration/CMakeLists.txt
@@ -24,5 +24,5 @@ target_include_directories(wls_integration_lib PRIVATE include/)
 target_compile_definitions(wls_integration_lib PUBLIC ENABLE_WLS)
 target_link_libraries(wls_integration_lib PRIVATE nfapi_nr_lib)
 target_link_libraries(wls_integration_lib PRIVATE nr_fapi_p5 nr_fapi_p7)
-target_link_libraries(wls_integration_lib PRIVATE nfapi_vnf_lib nfapi_pnf_lib)
+target_link_libraries(wls_integration_lib PRIVATE nfapi_lte_vnf_lib nfapi_nr_vnf_lib nfapi_pnf_lib)
 target_link_libraries(wls_integration_lib PRIVATE "${WLS_LIB}" "${hugetlbfs_LIB}")
